requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
djangorestframework==3.12.4
Faker==9.7.1
gunicorn==20.1.0
kombu==5.1.0
prompt-toolkit==3.0.21
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
whitenoise==5.3.0
